pyproject.toml
[project]
name = "ragger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Fabrizio Damicelli", email = "fabridamicelli@gmail.com" }
]
requires-python = ">=3.12"

dependencies = [
  "fastapi>=0.116.1",
  "python-dotenv>=1.1.1",
  "qdrant-client>=1.15.1",
  "requests>=2.32.5",
  "sentence-transformers>=5.1.0",
  "torch>=2.7.0",
  "torchvision>=0.22.0",
  "uvicorn>=0.35.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true


[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"
